- [fix] removed internal imports on 'mod.ts' from within the lib - [feat] removed encoded entry, lib will only export byte array oriented kv, moved kv types to types, and exported them from the internal_mod
